Add script to fix git add --patch with diff-so-fancy
This is a hacky workaround for the issue below, but it works with all
the test cases I could come up with so far. I'll try to get a perl
version of this fix merged upstream soon, but need to sit with this for
a bit until I'm fully convinced all the various edge cases work.

https://github.com/so-fancy/diff-so-fancy/issues/35