Enable and configure vim-visual-multi for multiple cursors
This also required reconfiguring vim-sandwich to use vim-surround's
keybindings.

This happened for two reasons:

1. vim-sandwich's default keybinding uses the `s` key (does same thing as
   `cl` in vim's default keybinding). This confuses and angers some
   plugins (and people).

2. vim-visual-multi does not seem to consistently detect user-defined
   keybindings in normal mode. For example, from vim-sandwich's default
   keybindings, `sa` doesn't work in visual-multi, but `sd` does (even
   though they both start with the overridden `s` key).

Either way, vim-visual-multi's maintainer (quite understandably) doesn't
want to support every plugin. And the related issue for this
(mg979/vim-visual-multi#128) was closed as "won't fix" so we're on our
own to work it out if we want to keep using it (which I do because it's
the best multiple cursor plugin available right now, and this is one of
the last features I still miss from my sublime text days).

@todo maybe file an issue about the inconsistency in keybindings?